+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY5ZBRJNE0XD5YMW1PRX0P7X"
+        "01KY6CT7C7Q1NPT11PS0QEVH5F"
     )
     r"""The unique identifier of the data source"""