+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -55,9 +55,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-22T22:32:40.794Z"
+    created: Optional[str] = "2026-07-23T02:27:42.927Z"
 
-    updated: Optional[str] = "2026-07-22T22:32:40.794Z"
+    updated: Optional[str] = "2026-07-23T02:27:42.927Z"
 
     updated_by_id: OptionalNullable[str] = UNSET