+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY5ZBRJGRRGNW5YSESK3JDJJ"
+        "01KY6CT7C0EQMXP92QY35459M9"
     )
     r"""The unique identifier of the data source"""