+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY5ZBRJKRQ6JRPHPYT4J147P"
+        "01KY6CT7C40FZ3DJJ2JRJFCJYY"
     )
     r"""The unique identifier of the data source"""