+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY5ZBRJRXRF2VA4TRJZ9JJSD"
+        "01KY6CT7CC998TG5PVTKF9H2DC"
     )
     r"""The unique identifier of the data source"""