+++ packages/orq-rc/RELEASES.md
@@ -9689,3 +9689,13 @@ Based on:
 - [python v4.13.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.0 - packages/orq-rc
+
+## 2026-07-23 02:30:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.1 - packages/orq-rc